Moved QuestionAnswerHTML away from container logic
